+++ postgres/src/transaction.rs
@@ -1,12 +1,10 @@
 mod builder;
 mod portal;
 
-use core::ops::{Deref, DerefMut};
-
 use std::borrow::Cow;
 
 use super::{
-    client::{Client, ClientBorrowMut},
+    client::{Client, ClientBorrow, ClientBorrowMut},
     driver::codec::AsParams,
     error::Error,
     execute::Execute,
@@ -18,90 +16,84 @@ use super::{
 pub use builder::{IsolationLevel, TransactionBuilder};
 pub use portal::Portal;
 
-pub struct Transaction<'a, C>
-where
-    C: ClientBorrowMut,
-{
-    client: MaybeOwnedClient<'a, C>,
-    save_point: SavePoint,
+struct SavePoint {
+    name: Name,
     state: State,
 }
 
-enum MaybeOwnedClient<'a, C> {
-    Owned(C),
-    Borrowed(&'a mut C),
-}
-
-impl<C> MaybeOwnedClient<'_, C> {
-    fn reborrow<'a>(&'a mut self) -> MaybeOwnedClient<'a, C> {
-        match self {
-            Self::Borrowed(cli) => MaybeOwnedClient::Borrowed(cli),
-            Self::Owned(cli) => MaybeOwnedClient::Borrowed(cli),
-        }
-    }
-}
-
-impl<C> Deref for MaybeOwnedClient<'_, C> {
-    type Target = C;
-
-    fn deref(&self) -> &Self::Target {
-        match self {
-            Self::Owned(cli) => cli,
-            Self::Borrowed(cli) => cli,
-        }
-    }
-}
-
-impl<C> DerefMut for MaybeOwnedClient<'_, C> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        match self {
-            Self::Owned(cli) => cli,
-            Self::Borrowed(cli) => cli,
-        }
-    }
-}
-
-enum SavePoint {
+enum Name {
     None,
     Auto { depth: u32 },
     Custom { name: String, depth: u32 },
 }
 
+impl Default for SavePoint {
+    fn default() -> Self {
+        Self {
+            name: Name::None,
+            state: State::WantRollback,
+        }
+    }
+}
+
 impl SavePoint {
-    fn nest_save_point(&self, name: Option<String>) -> Self {
-        match *self {
-            Self::None => match name {
-                Some(name) => SavePoint::Custom { name, depth: 1 },
-                None => SavePoint::Auto { depth: 1 },
+    // rollback runs in Drop trait impl. the execution part has to run eargerly
+    fn rollback(&mut self, cli: impl ClientBorrowMut) -> impl Future<Output = Result<(), Error>> + Send {
+        self.state = State::Finish;
+
+        let fut = match self.name {
+            Name::None => Cow::Borrowed("ROLLBACK"),
+            Name::Auto { depth } => Cow::Owned(format!("ROLLBACK TO sp_{depth}")),
+            Name::Custom { ref name, .. } => Cow::Owned(format!("ROLLBACK TO {name}")),
+        }
+        .execute(cli.borrow_cli_ref());
+
+        async { fut.await.map(|_| ()) }
+    }
+
+    async fn commit(&mut self, cli: impl ClientBorrowMut) -> Result<(), Error> {
+        self.state = State::Finish;
+
+        match self.name {
+            Name::None => Cow::Borrowed("COMMIT"),
+            Name::Auto { depth } => Cow::Owned(format!("RELEASE sp_{depth}")),
+            Name::Custom { ref name, .. } => Cow::Owned(format!("RELEASE {name}")),
+        }
+        .execute(cli.borrow_cli_ref())
+        .await?;
+
+        Ok(())
+    }
+
+    async fn nest_save_point(&self, cli: impl ClientBorrowMut, name: Option<String>) -> Result<Self, Error> {
+        let name = match self.name {
+            Name::None => match name {
+                Some(name) => Name::Custom { name, depth: 1 },
+                None => Name::Auto { depth: 1 },
             },
-            Self::Auto { depth } | Self::Custom { depth, .. } => match name {
-                Some(name) => SavePoint::Custom { name, depth },
-                None => SavePoint::Auto { depth: depth + 1 },
+            Name::Auto { depth } | Name::Custom { depth, .. } => match name {
+                Some(name) => Name::Custom { name, depth },
+                None => Name::Auto { depth: depth + 1 },
             },
+        };
+
+        match self.name {
+            Name::None => Cow::Borrowed("SAVEPOINT"),
+            Name::Auto { depth } => Cow::Owned(format!("SAVEPOINT sp_{depth}")),
+            Name::Custom { ref name, .. } => Cow::Owned(format!("SAVEPOINT {name}")),
         }
+        .execute(cli.borrow_cli_ref())
+        .await?;
+
+        Ok(SavePoint {
+            name,
+            state: State::WantRollback,
+        })
     }
 
-    fn save_point_query(&self) -> Cow<'static, str> {
-        match self {
-            Self::None => Cow::Borrowed("SAVEPOINT"),
-            Self::Auto { depth } => Cow::Owned(format!("SAVEPOINT sp_{depth}")),
-            Self::Custom { name, .. } => Cow::Owned(format!("SAVEPOINT {name}")),
-        }
-    }
-
-    fn commit_query(&self) -> Cow<'static, str> {
-        match self {
-            Self::None => Cow::Borrowed("COMMIT"),
-            Self::Auto { depth } => Cow::Owned(format!("RELEASE sp_{depth}")),
-            Self::Custom { name, .. } => Cow::Owned(format!("RELEASE {name}")),
-        }
-    }
-
-    fn rollback_query(&self) -> Cow<'static, str> {
-        match self {
-            Self::None => Cow::Borrowed("ROLLBACK"),
-            Self::Auto { depth } => Cow::Owned(format!("ROLLBACK TO sp_{depth}")),
-            Self::Custom { name, .. } => Cow::Owned(format!("ROLLBACK TO {name}")),
+    fn on_drop(&mut self, cli: impl ClientBorrowMut) {
+        if matches!(self.state, State::WantRollback) {
+            drop(self.rollback(cli));
         }
     }
 }
@@ -111,15 +103,20 @@ enum State {
     Finish,
 }
 
+pub struct Transaction<'a, C>
+where
+    C: ClientBorrowMut,
+{
+    client: &'a mut C,
+    save_point: SavePoint,
+}
+
 impl<C> Drop for Transaction<'_, C>
 where
     C: ClientBorrowMut,
 {
     fn drop(&mut self) {
-        match self.state {
-            State::WantRollback => self.do_rollback(),
-            State::Finish => {}
-        }
+        self.save_point.on_drop(&mut self.client);
     }
 }
 
@@ -166,61 +163,115 @@ where
 
     /// Consumes the transaction, committing all changes made within it.
     pub async fn commit(mut self) -> Result<(), Error> {
-        self.state = State::Finish;
-        self.save_point
-            .commit_query()
-            .execute(self.client.borrow_cli_ref())
-            .await?;
-        Ok(())
+        self.save_point.commit(&mut self.client).await
     }
 
     /// Rolls the transaction back, discarding all changes made within it.
     ///
     /// This is equivalent to [`Transaction`]'s [`Drop`] implementation, but provides any error encountered to the caller.
     pub async fn rollback(mut self) -> Result<(), Error> {
-        self.state = State::Finish;
-        self.save_point
-            .rollback_query()
-            .execute(self.client.borrow_cli_ref())
-            .await?;
-        Ok(())
+        self.save_point.rollback(&mut self.client).await
     }
 
-    fn new_owned<'a>(client: C) -> Transaction<'a, C>
-    where
-        C: 'a,
-    {
+    fn new(client: &mut C) -> Transaction<'_, C> {
         Transaction {
-            client: MaybeOwnedClient::Owned(client),
-            save_point: SavePoint::None,
-            state: State::WantRollback,
-        }
-    }
-
-    fn new_ref(client: &mut C) -> Transaction<'_, C> {
-        Transaction {
-            client: MaybeOwnedClient::Borrowed(client),
-            save_point: SavePoint::None,
-            state: State::WantRollback,
+            client,
+            save_point: SavePoint::default(),
         }
     }
 
     async fn _save_point(&mut self, name: Option<String>) -> Result<Transaction<'_, C>, Error> {
-        let save_point = self.save_point.nest_save_point(name);
-        save_point
-            .save_point_query()
-            .execute(self.client.borrow_cli_ref())
-            .await?;
-
+        let save_point = self.save_point.nest_save_point(&mut self.client, name).await?;
         Ok(Transaction {
-            client: self.client.reborrow(),
+            client: self.client,
             save_point,
-            state: State::WantRollback,
         })
     }
+}
 
-    fn do_rollback(&mut self) {
-        drop(self.save_point.rollback_query().execute(self.client.borrow_cli_ref()));
+pub struct TransactionOwned<C>
+where
+    C: ClientBorrowMut,
+{
+    client: C,
+    save_point: SavePoint,
+}
+
+impl<C> Drop for TransactionOwned<C>
+where
+    C: ClientBorrowMut,
+{
+    fn drop(&mut self) {
+        self.save_point.on_drop(&mut self.client);
+    }
+}
+
+impl<C> TransactionOwned<C>
+where
+    C: ClientBorrowMut,
+{
+    /// Binds a statement to a set of parameters, creating a [`Portal`] which can be incrementally queried.
+    ///
+    /// Portals only last for the duration of the transaction in which they are created, and can only be used on the
+    /// connection that created them.
+    pub async fn bind<'p>(
+        &'p self,
+        statement: &'p Statement,
+        params: &[&(dyn ToSql + Sync)],
+    ) -> Result<Portal<'p, C>, Error> {
+        self.bind_raw(statement, params.iter().cloned()).await
+    }
+
+    /// A maximally flexible version of [`Transaction::bind`].
+    pub async fn bind_raw<'p, I>(&'p self, statement: &'p Statement, params: I) -> Result<Portal<'p, C>, Error>
+    where
+        I: AsParams,
+    {
+        Portal::new(&self.client, statement, params).await
+    }
+
+    /// Like [`Client::transaction`], but creates a nested transaction via a savepoint.
+    ///     
+    /// [`Client::transaction`]: crate::client::Client::transaction
+    pub async fn transaction(&mut self) -> Result<Transaction<'_, C>, Error> {
+        self._save_point(None).await
+    }
+
+    /// Like [`Client::transaction`], but creates a nested transaction via a savepoint with the specified name.
+    ///
+    /// [`Client::transaction`]: crate::client::Client::transaction
+    pub async fn save_point<I>(&mut self, name: I) -> Result<Transaction<'_, C>, Error>
+    where
+        I: Into<String>,
+    {
+        self._save_point(Some(name.into())).await
+    }
+
+    /// Consumes the transaction, committing all changes made within it.
+    pub async fn commit(mut self) -> Result<(), Error> {
+        self.save_point.commit(&mut self.client).await
+    }
+
+    /// Rolls the transaction back, discarding all changes made within it.
+    ///
+    /// This is equivalent to [`Transaction`]'s [`Drop`] implementation, but provides any error encountered to the caller.
+    pub async fn rollback(mut self) -> Result<(), Error> {
+        self.save_point.rollback(&mut self.client).await
+    }
+
+    fn new(client: C) -> Self {
+        TransactionOwned {
+            client,
+            save_point: SavePoint::default(),
+        }
+    }
+
+    async fn _save_point(&mut self, name: Option<String>) -> Result<Transaction<'_, C>, Error> {
+        let save_point = self.save_point.nest_save_point(&mut self.client, name).await?;
+        Ok(Transaction {
+            client: &mut self.client,
+            save_point,
+        })
     }
 }
 
@@ -259,5 +310,43 @@ where
     #[inline]
     fn query(self, cli: &'c mut Transaction<PoolConnection<'p>>) -> Self::QueryOutput {
         Q::query(self, &mut *cli.client)
+    }
+}
+
+impl<'c, C, Q, EO, QO> Execute<&'c TransactionOwned<C>> for Q
+where
+    C: ClientBorrowMut,
+    Q: Execute<&'c Client, ExecuteOutput = EO, QueryOutput = QO>,
+{
+    type ExecuteOutput = EO;
+    type QueryOutput = QO;
+
+    #[inline]
+    fn execute(self, cli: &'c TransactionOwned<C>) -> Self::ExecuteOutput {
+        Q::execute(self, cli.client.borrow_cli_ref())
+    }
+
+    #[inline]
+    fn query(self, cli: &'c TransactionOwned<C>) -> Self::QueryOutput {
+        Q::query(self, cli.client.borrow_cli_ref())
+    }
+}
+
+// special treatment for pool connection for it's internal caching logic that are not accessible through Query trait
+impl<'c, 'p, Q, EO, QO> Execute<&'c mut TransactionOwned<PoolConnection<'p>>> for Q
+where
+    Q: Execute<&'c mut PoolConnection<'p>, ExecuteOutput = EO, QueryOutput = QO>,
+{
+    type ExecuteOutput = EO;
+    type QueryOutput = QO;
+
+    #[inline]
+    fn execute(self, cli: &'c mut TransactionOwned<PoolConnection<'p>>) -> Self::ExecuteOutput {
+        Q::execute(self, &mut cli.client)
+    }
+
+    #[inline]
+    fn query(self, cli: &'c mut TransactionOwned<PoolConnection<'p>>) -> Self::QueryOutput {
+        Q::query(self, &mut cli.client)
     }
 }