+++ postgres/CHANGES.md
@@ -3,7 +3,8 @@
 - add `error::ClosedByDriver` error type
 - add `pool::PoolBuilder::cache_size` for configuring the max size of statement cache of connection pool
 - add `dev::ClientBorrow` trait for extending client type. It's used as a super trait of `ClientBorrowMut`
-- add `transaction::TransactionBuilder::begin_owned` method for specialized handling of transaction with owned client type 
+- add `transaction::TransactionBuilder::begin_owned` method for specialized handling of transaction with owned client type
+- add `transaction::TransactionOwned` type to complement `TransactionBulder::begin_owned` method
 
 ## Change
 - switch to `tokio-uring-xitca` for `io-uring` feature. IO uring driver must run in it's runtime instead of original `tokio-uring`